emis model parent class
